Add check_needless_collect to its own module
